Revert "Merge remote-tracking branch 'upstream/master' into AdvancedBachrusAutoSpeedMatcher"
This reverts commit 91ff402a1881ba869525b0ee043b69a7da53003a, reversing
changes made to e933601cde520901108ff1583e173b4ae6cc623b.

+++ java/src/jmri/jmrit/turnoutoperations/CommonTurnoutOperationConfig.java
@@ -22,7 +22,6 @@ public class CommonTurnoutOperationConfig extends TurnoutOperationConfig {
     /**
      * Create the config JPanel, if there is one, to configure this operation
      * type.
-     * @param op turnout operation.
      */
     public CommonTurnoutOperationConfig(TurnoutOperation op) {
         super(op);
@@ -68,9 +67,9 @@ public class CommonTurnoutOperationConfig extends TurnoutOperationConfig {
      */
     @Override
     public void endConfigure() {
-        int newInterval = ((Integer) intervalSpinner.getValue());
+        int newInterval = ((Integer) intervalSpinner.getValue()).intValue();
         myOp.setInterval(newInterval);
-        int newMaxTries = ((Integer) maxTriesSpinner.getValue());
+        int newMaxTries = ((Integer) maxTriesSpinner.getValue()).intValue();
         myOp.setMaxTries(newMaxTries);
     }
